New: converters.registrar module for outbound format registration
The Registrar class will be used to help Lychee determine which data types
to produce through outbound conversion. This will be used by the new
REGISTER_FORMAT and UNREGISTER_FORMAT signals, which will replace
WHO_IS_LISTENING and I_AM_LISTENING. While the latter is potentially easier,
it also won't work with an asynchronous connection to the user interface.