Refactor: document.Document now uses <meiHead> that's valid (L)MEI
Previously, the <meiHead> elements were just empty. That's not valid MEI,
and it's not valid Lychee-MEI either (that is, now that I looked up what
MEI-proper says about it...).

Now, Document uses the minimum required <meiHead> by default.

There is still no way to get/set *part* of the header---just the header
as a whole. This comes next.