Remove all the "document step" signals
We've decided that the "document step" does not need its own signals, because
replacing the 'document' module implies you're not using MEI, which implies
you must replace everything... and should therefore not be using Lychee. That
is, Lychee's 'document' module is too central to Lychee's operation that we
don't want third-party software interfering.

For T9