Removed most of the "VCS step" signals
I've now reduced the number of VCS signals to those required to run the
step itself. As long as the workflow order is the same, the signals are
now in line with what's required to abstract across a variety of VCSes.

So yes, somebody will be able to write a Git plugin.

Closes T9