Change: document._write_out() saves XML with pretty_print=True
Previously we were using C14N serialization, which is more verbose and only
uses a single line for the whole file. This made the output very hard to
read, and also meant that *any* change to a file would cause Hg to register
the *whole* file as different.

Now we don't use C14N, do use "pretty printing," and also add an XML
delcaration.