New: ACTION_START signal accepts "revision" kwarg
This new argument for ACTION_START allows requesting data from changesets
other than the most recent for an outbound-only action. The revision is
changed only temporarily, and not changed at all when there is an inbound
change to the score.

(This commit also introduces the 300th test to Lychee).

For T23