Change: ACTION_START's "views_info" used for inbound sect ID (part 1 of 2)
As required for T33, the "views_info" parameter for the ACTION_START signal
can now also be used to tell Lychee which section is coming inbound from an
external format, thus obviating the need for a "section cursor" in Lychee.

Part 1: modifications to the 'workflow' and 'steps' modules. The 'views'
modules will be modified next.

For T33