Fixed: InteractiveSession._action_start() can start from tagged revision
Previously, if the repo started on a tagged changeset, _action_start() would
include the rev number, hash, and tag in the request to Mercurial to update
the repo to the initial state, after doing the outbound computation. However,
this causes an error because Hg expects only rev number and hash. So now it
sends only the rev number and hash.

For T32