New: helper scripts for LMEI/LilyPond conversion
These four command-line scripts convert between LilyPond and LMEI document. They are for use by developers working on Lychee converters to check for unintended information loss and other discrepancies during the conversion process. Like other helper scripts, they do not require going through a Lychee action for use.

One of the scripts, a round-trip LMEI-to-LilyPond-to-LMEI converter, has a "--check" option that throws an error if the output XML is different from the input with the exception of xml:id attributes.

To reduce duplicate code, a layer of abstraction is added (in helper_utils.py) that takes care of creating the argument parser and reading and writing to the files.

For T133.