Tests: fixed for Document.save_everything()
Before VCS disablement, my use of a Session instance in the Document class's
tests meant that a new Hg repo was created and destroyed with many of the
tests. It also meant the save_everything() tests checked for the presence
of the ".hg" directory. Now that VCS is disabled by default, and disabled
in these tests, the ".hg" directory isn't there, which caused those tests
to fail.

So I fixed the tests.

For T111