Refactor: inbound LilyPond rests, spacers, and dots (WIP)
After changing the Grako grammar, the document structure given to the inbound
LilyPond converter is quite different and must be refactored. This commits
starts with the rest, spacer rest, and dot imports, adds appropriate
documentation, and also appropriate tests.

In addition, this commit is an experiment with the Lithoxyl logging library,
which we are adopting instead of the standard library's "logging" module.

For T85, T96