Refactor: the LilyPond EBNF grammar to mark "staff_content" as "staff"
Previously the "staff" production included "staff_content," but only the
former added the "ly_type" member. In terms of converting to LMEI, it doesn't
matter whether the extra \new Staff { } stuff (of the "staff" production)
is present, or it's just the staff content. Therefore, both productions
now include the "ly_type" member, and can go through the translation step
identically.

For T85