Refactor: lychee.workflow.steps to use Lithoxyl
Previously, the "steps" module used the lychee.log() function. Now it uses
Lithoxyl. We'll have to revisit the instrumentation/logging in the future
as we adjust error handling and workflow management, but for now, each of
the workflow steps and many of the sub-steps are instrumented, so it's
not too bad.

For T96