Refactor: Remove document.MOVE_SECTION_TO signal
This was the only signal in the "document" signals, so I removed that module
too. For some reason, there were two move_section_to() functions in the
"document" module proper, so I also removed the module-leve function which
leaves the only on the Document class, which also happens to be tested.

For T108