Change: better top-level LY parsing, support "\new Score", "\language"
This commit improves parsing of top-level structures in LilyPond files to support a wider subset of the LilyPond grammar, and to exclude invalid constructs that were previously allowed.

First, "bare" music at the top level is now disallowed. It is invalid LilyPond and shouldn't be allowed by Lychee either.

"\new Score" is now allowed in place of "\score". This is important since Abjad outputs "\new Score".

"\version" statements are now "free-floating", instead of being restricted to immediately before a score as before. They are also optional as in LilyPond.

The "\language" command is now parsed, but not interpreted.