Add: support \score { \new Score << ... >> }
This commit adds support in the inbound LilyPond parser for parsing a score nested within a score, a construct sometimes produced by Abjad. The solution is a quick fix to allow supporting multiple staves in Abjad. In the future, we will want the parser and inbound converter arbitrary trees of contexts (within reason).