Add: support for user settings and outbound LY language
This commit adds support for setting the language in the outbound LilyPond converter. This is achieved by implementing a new feature called "user settings," which introduces a new file user/lychee.settings.xml in the Lychee repo directory that contains settings specific to the user, which are not committed to VCS. A convenient interface is provided where inbound and outbound converters are passed a dictionary called "user_settings," converted to and from XML via xmltodict. Coverters can read from and write to this dictionary to interact with user settings.