Replace Grako with TatSu
Grako has been superseded by TatSu. TatSu is mostly backward compatible
with Grako; the differences between them do not affect Lychee, so this
commit simply replaces "grako" with "tatsu" everywhere.

There was also a fragile if statement in
lychee.converters.inbound.do_document that incorrectly used type()
instead of isinstance() that broke when moving to TatSu. This was also
fixed in the process.