Change: InteractiveSession._cleanup_for_new_action() to delete editor views
When cleaning up before an incoming change, InteractiveSession will check
for existing saved text editor contents for that <section>. If any exist,
they will be deleted to guarantee the outbound converter produces results
that are consistent with the state of the LMEI.

This assumes that an inbound change will be followed by re-saving the editor
contents, but that is left as an exercise for the reader. I mean, Julius
can do that.

For T194